Add tests/e2e/scale/test_osd_node_balancing.py (#4133)
This test checks osd distributions for 4.7 clusters where 3 nodes and 3 osds
are added at a time. This should work for both flexible scaling and non-flexible
situations. This makes sure that one case of OCSQE-571 is tested, and also makes
sure that comment #33 in bugzilla Bug 1817438 is checked. This tests keeps track
of information collected in the elastic search database.

Signed-off-by: wusui <wusui@redhat.com>